fix: 修复首次云端构建的四簇编译错误
- HistoryActivity: 补 android.view.View import
- ShellSkins.CutCornerDrawable.buildPath: Path.lineTo 的 Int 坐标补 toFloat()（8 处，Kotlin 无 Int→Float 隐转）
- TextCollector: dfs 递归调用漏传 excludePackage——面板自抓免疫在递归层失效的功能性 bug
- DsFx.BeamView: 0x801EA5C7 超 Int 范围，补 toInt()

+++ app/src/main/java/com/speedtrans/app/ui/DsFx.kt
@@ -71,7 +71,7 @@ class BeamView @JvmOverloads constructor(
         this.w = w; this.h = h
         if (w <= 0) return
         beam.shader = LinearGradient(0f, 0f, w.toFloat(), 0f,
-            intArrayOf(0x001EA5C7, 0x801EA5C7, 0x001EA5C7),
+            intArrayOf(0x001EA5C7, 0x801EA5C7.toInt(), 0x001EA5C7),
             floatArrayOf(0f, 0.5f, 1f), Shader.TileMode.CLAMP)
         glow.shader = LinearGradient(0f, 0f, w.toFloat(), 0f,
             intArrayOf(0x001EA5C7, 0x201EA5C7, 0x001EA5C7),
